backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
